# Improve comparison performance