--- a/src/FlowtideDotNet.Core/Properties/AssemblyInfo.cs
+++ b/src/FlowtideDotNet.Core/Properties/AssemblyInfo.cs
@@ -17,3 +17,4 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("FlowtideDotNet.Connector.SpiceDB")]
 [assembly: InternalsVisibleTo("FlowtideDotNet.Connector.Permify")]
 [assembly: InternalsVisibleTo("FlowtideDotNet.Connector.Sharepoint")]
+[assembly: InternalsVisibleTo("FlowtideDotNet.Benchmarks")]